gfx::frame: Change vertex coords to int
This change has three main benefits:

 1. Solves any rounding issues when handling or exporting vertex
 coordinates, for example when exporting WGEOv2 entries.

 2. Removes the requirement for the GL_ARB_shader_bit_encoding OpenGL
 extension. This also helps with an issue on certain Sandy Bridge Intel
 HD Graphics driver versions on Windows.

 3. Allows for more complicated bitwise and integer overflow behaviors
 which will likely be needed to implement compressed object model
 animations used in Crash 2 and Crash 3.

+++ src/edit.hh
@@ -1170,9 +1170,9 @@ private:
 
     // (var) m_x_field, m_y_field, m_z_field
     // The subfields for X/Y/Z vertex components.
-    field<float> m_x_field;
-    field<float> m_y_field;
-    field<float> m_z_field;
+    field<int> m_x_field;
+    field<int> m_y_field;
+    field<int> m_z_field;
 
     // (var) m_fx_field
     // The subfield for the vertex special effects bits.
@@ -1188,17 +1188,17 @@ public:
     // the event on the outer field (this).
     field()
     {
-        m_x_field.on_change <<= [this](float new_value) {
+        m_x_field.on_change <<= [this](int new_value) {
             auto new_vtx = *m_object;
             new_vtx.x = new_value;
             on_change(new_vtx);
         };
-        m_y_field.on_change <<= [this](float new_value) {
+        m_y_field.on_change <<= [this](int new_value) {
             auto new_vtx = *m_object;
             new_vtx.y = new_value;
             on_change(new_vtx);
         };
-        m_z_field.on_change <<= [this](float new_value) {
+        m_z_field.on_change <<= [this](int new_value) {
             auto new_vtx = *m_object;
             new_vtx.z = new_value;
             on_change(new_vtx);
